+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState/Types.hs
@@ -95,6 +95,17 @@ instance DecCBOR AccountState where
   decCBOR =
     decodeRecordNamed "AccountState" (const 2) $ AccountState <$> decCBOR <*> decCBOR
 
+instance ToJSON AccountState where
+  toJSON = object . toAccountStatePairs
+  toEncoding = pairs . mconcat . toAccountStatePairs
+
+toAccountStatePairs :: KeyValue a => AccountState -> [a]
+toAccountStatePairs as@(AccountState _ _) =
+  let AccountState {asTreasury, asReserves} = as
+   in [ "treasury" .= asTreasury
+      , "reserves" .= asReserves
+      ]
+
 instance NoThunks AccountState
 
 instance NFData AccountState
@@ -174,6 +185,27 @@ instance (EraTxOut era, EraGovernance era) => ToCBOR (EpochState era) where
 
 instance (EraTxOut era, EraGovernance era) => FromCBOR (EpochState era) where
   fromCBOR = fromEraCBOR @era
+
+instance (EraTxOut era, EraGovernance era) => ToJSON (EpochState era) where
+  toJSON = object . toEpochStatePairs
+  toEncoding = pairs . mconcat . toEpochStatePairs
+
+toEpochStatePairs ::
+  ( EraTxOut era
+  , EraGovernance era
+  , KeyValue a
+  ) =>
+  EpochState era ->
+  [a]
+toEpochStatePairs es@(EpochState _ _ _ _ _ _) =
+  let EpochState {..} = es
+   in [ "esAccountState" .= esAccountState
+      , "esSnapshots" .= esSnapshots
+      , "esLState" .= esLState
+      , "esPrevPp" .= esPrevPp
+      , "esPp" .= esPp
+      , "esNonMyopic" .= esNonMyopic
+      ]
 
 -- =============================
 
@@ -302,6 +334,21 @@ instance (EraTxOut era, EraGovernance era) => ToCBOR (UTxOState era) where
 
 instance (EraTxOut era, EraGovernance era) => FromCBOR (UTxOState era) where
   fromCBOR = toPlainDecoder (eraProtVerLow @era) decNoShareCBOR
+
+instance (EraTxOut era, EraGovernance era) => ToJSON (UTxOState era) where
+  toJSON = object . toUTxOStatePairs
+  toEncoding = pairs . mconcat . toUTxOStatePairs
+
+toUTxOStatePairs ::
+  (EraTxOut era, EraGovernance era, KeyValue a) => UTxOState era -> [a]
+toUTxOStatePairs utxoState@(UTxOState _ _ _ _ _) =
+  let UTxOState {..} = utxoState
+   in [ "utxo" .= utxosUtxo
+      , "deposited" .= utxosDeposited
+      , "fees" .= utxosFees
+      , "ppups" .= utxosGovernance
+      , "stake" .= utxosStakeDistr
+      ]
 
 -- | New Epoch state and environment
 data NewEpochState era = NewEpochState
@@ -476,6 +523,18 @@ instance (EraTxOut era, EraGovernance era) => ToCBOR (LedgerState era) where
 
 instance (EraTxOut era, EraGovernance era) => FromCBOR (LedgerState era) where
   fromCBOR = toPlainDecoder (eraProtVerLow @era) decNoShareCBOR
+
+instance (EraTxOut era, EraGovernance era) => ToJSON (LedgerState era) where
+  toJSON = object . toLedgerStatePairs
+  toEncoding = pairs . mconcat . toLedgerStatePairs
+
+toLedgerStatePairs ::
+  (EraTxOut era, EraGovernance era, KeyValue a) => LedgerState era -> [a]
+toLedgerStatePairs ls@(LedgerState _ _) =
+  let LedgerState {..} = ls
+   in [ "utxoState" .= lsUTxOState
+      , "delegationState" .= lsDPState
+      ]
 
 -- ====================================================